+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 * Requires go 1.18 or higher
 * The "interface{}" type replaced by "any"
-* Added "accent-color", "overflow", "arrow", "arrow-align", "arrow-size", "arrow-width", and "arrow-offset" properties 
+* Added "accent-color", "tab-size", "overflow", "arrow", "arrow-align", "arrow-size", "arrow-width", and "arrow-offset" properties 
 * Added "@ruiArrowSize" and "@ruiArrowWidth" constants to the default theme
 * Added Transition, Transitions, and SetTransition functions to the ViewStyle interface
-* Added GetAccentColor, GetOverflow, IsTimingFunctionValid, and GetTransitions functions
+* Added GetAccentColor, GetTabSize, GetOverflow, IsTimingFunctionValid, and GetTransitions functions
 * Changed GetTransition functions
 * Added the OpenURL function to the Session interface